% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_Index.R
\name{B_index}
\alias{B_index}
\title{B index a la Nonacs.}
\usage{
B_index(r, t)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of exposure times.}
}
\value{
The B index.
}
\description{
B index a la Nonacs.
}
\examples{
set.seed(1)
RS = rpois(100, 5) 
Age = rpois(100, 45)
B_index(RS, Age)
}
