% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{KipsigisMales}
\alias{KipsigisMales}
\title{KipsigisMales}
\format{A data frame with 848 rows and 3 variables:
\describe{
  \item{\code{code}}{Personal ID code.}
  \item{\code{age}}{Age at death or last census.}
  \item{\code{rs}}{Reproductive success.}
}}
\source{
Monique Borgerhoff Mulder
}
\usage{
KipsigisMales
}
\description{
Reproductive success and age for a census of Kipsigis men.
}
\keyword{datasets}
