% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Index.R
\name{M_index}
\alias{M_index}
\title{M index of reproductive skew.}
\usage{
M_index(r, t, Samples = 1000)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of exposure times.}

\item{Samples}{Number of samples used to estimate the expected value of the correction term.}
}
\value{
The M index.
}
\description{
M index of reproductive skew.
}
\examples{
set.seed(1) 
RS = rpois(100, 5) 
Age = rpois(100, 45)
M_index(RS, Age)
}
