% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Index_Age.R
\name{M_index_age}
\alias{M_index_age}
\title{M index accounting for diminishing returns to age.}
\usage{
M_index_age(r, t, t0 = 0)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of ages at death or last census.}

\item{t}{A vector of ages at first census. Defaults to zero.}
}
\value{
The M index accounting for diminishing returns to age.
}
\description{
M index accounting for diminishing returns to age.
}
\examples{
RS = rpois(100, 5) 
Age = rpois(100, 45)
M_index_age(RS, Age)
}
