% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Index_Stan.R
\name{M_index_stan}
\alias{M_index_stan}
\title{M index of reproductive skew using Stan to simulate the population distribution of RS and age.}
\usage{
M_index_stan(r, t, t0 = FALSE, Samples = 2000, Warmup = 1000,
  Chains = 1, adapt_delta = 0.9, max_treedepth = 12)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of ages at death or last census.}

\item{t}{A vector of ages at first census. Defaults to zero.}
}
\value{
A Stan object is saved to the workspace. This will be used for plots and calculations. The M and Mraw results have 4 columns each:

1) The point estimate on the sample data.

2) The point estimate on the sample data, but adjusting for diminising fitness returns to age.

3) The posterior estimate using data simulated from the generative model.

4) The posterior estimate using data simulated from the generative model, but adjusting for diminising fitness returns to age.
}
\description{
M index of reproductive skew using Stan to simulate the population distribution of RS and age.
}
\examples{
RS = rpois(100, 5) 
Age = rpois(100, 45)
M_index_stan(RS, Age)
}
