% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mraw_Index.R
\name{Mraw_index}
\alias{Mraw_index}
\title{Mraw index of reproductive skew.}
\usage{
Mraw_index(r, t)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of exposure times.}
}
\value{
The Mraw index.
}
\description{
Mraw index of reproductive skew.
}
\examples{
RS = rpois(100, 5) 
Age = rpois(100, 45)
Mraw_index(RS, Age)
}
