% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mraw_Index_Age.R
\name{Mraw_index_age}
\alias{Mraw_index_age}
\title{Mraw index accounting for diminishing returns to age.}
\usage{
Mraw_index_age(r, t, t0 = 0)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of ages at death or last census.}

\item{t}{A vector of ages at first census. Defaults to zero.}
}
\value{
The Mraw index accounting for diminishing returns to age.
}
\description{
Mraw index accounting for diminishing returns to age.
}
\examples{
set.seed(1)
RS = rpois(100, 5) 
Age = rpois(100, 45)
Mraw_index_age(RS, Age)
}
