% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mraw_Index_from_B_Index.R
\name{Mraw_index_from_B_index}
\alias{Mraw_index_from_B_index}
\title{Calculate the Mraw index from the B index and relevant data.}
\usage{
Mraw_index_from_B_index(B, R, N)
}
\arguments{
\item{B}{The B index value estimated from a dataset.}

\item{R}{Total number of RS events in the sample from which B was estimated.}

\item{N}{Total number of individuals in the sample from which B was estimated.}
}
\value{
The Mraw index.
}
\description{
Calculate the Mraw index from the B index and relevant data.
}
\examples{
B = 0.1 
R = 50
N = 10
Mraw_index_from_B_index(B, R, N)
}
