% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCI.R
\name{PCI}
\alias{PCI}
\title{PCI - Stolen from rethinking by R. McElreath}
\usage{
PCI(samples, prob = 0.89)
}
\arguments{
\item{samples}{A vector of data-points.}

\item{prob}{A percentile interval.}
}
\value{
The equal tail credible interval.
}
\description{
PCI - Stolen from rethinking by R. McElreath
}
\examples{
set.seed(1)
RS = rpois(100, 5) 
PCI(RS, 0.9)
}
