% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skew_Diagnostics_Plot.R
\name{skew_diagnostics_plot}
\alias{skew_diagnostics_plot}
\title{Check the fit of the predictive model to the sample data}
\usage{
skew_diagnostics_plot(RS, T1, T0 = 0, SkewResults = StanReults,
  save = FALSE)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of ages at death or last census.}

\item{t}{A vector of ages at first census. Defaults to zero.}

\item{Save}{If desired, the plot can be save. Just set Save="FigueNameDesired", and R will export the plot as "FigureNameDesired.pdf".}
}
\description{
Check the fit of the predictive model to the sample data
}
\examples{
set.seed(1)
RS = rpois(100, 5) 
Age = rpois(100, 45)
M_index_stan(RS, Age)
skew_index_plot(Index="M", Age=FALSE, Save=FALSE)
}
