% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skew_Index_Plot.R
\name{skew_index_plot}
\alias{skew_index_plot}
\title{Simple posterior density plot for Stan results. 95 percent CI of posterior preditionsc in shaded orange, posterior median in white, and point estimate from sample in dark orange.}
\usage{
skew_index_plot(Index = "M", Age = FALSE, SkewResults = StanReults,
  Save = FALSE)
}
\arguments{
\item{Index}{Should plot be of "M" or of "Mraw"?}

\item{Save}{If desired, the plot can be save. Just set Save="FigueNameDesired", and R will export the plot as "FigureNameDesired.pdf".}

\item{t}{M or Mraw should be adjusted for diminishing returns to age, TRUE or FALSE?}
}
\description{
Simple posterior density plot for Stan results. 95 percent CI of posterior preditionsc in shaded orange, posterior median in white, and point estimate from sample in dark orange.
}
\examples{
set.seed(1)
RS = rpois(100, 5) 
Age = rpois(100, 45)
M_index_stan(RS, Age)
skew_index_plot(Index="M", Age=FALSE, Save=FALSE)
}
