% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{package}
\name{SkewCalc}
\alias{SkewCalc}
\alias{SkewCalc-package}
\title{SkewCalc: Estimate reproductive skew using Multinomial Index}
\description{
Documentation here is bare-bones, but will be updated as needed.
}
