% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elast.R
\name{elast}
\alias{elast}
\title{Estimate the elasticity of RS on age.}
\usage{
elast(r, t, t0 = 0)
}
\arguments{
\item{r}{A vector of RS values.}

\item{t}{A vector of ages at death or last census.}

\item{t0}{A vector of ages at first census. Defaults to zero.}
}
\value{
MLE of the elasticity of RS on Age, using Poisson regression.
}
\description{
Estimate the elasticity of RS on age.
}
\examples{
set.seed(1)
Age = rpois(100, 45)
RS = rep(NA,100)
for(i in 1:100) RS[i] = rpois(1, 5*Age^0.75) 
elast(RS, Age)
}
