% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Index_from_B_Index.R
\name{M_index_from_B_index}
\alias{M_index_from_B_index}
\title{Calculate the M index from the B index and relevant data.}
\usage{
M_index_from_B_index(B, R, N)
}
\arguments{
\item{B}{The B index value estimated from a dataset.}

\item{R}{Total number of RS events in the sample from which B was estimated.}

\item{R}{Total number of individuals in the sample from which B was estimated.}
}
\value{
The M index.
}
\description{
Calculate the M index from the B index and relevant data.
}
\examples{
B = 0.1 
R = 50
N = 10
M_index_from_B_index(B, R, N)
}
